(Triggers) Move triggers to separate submodule
