+++ sensormesh/application.py
@@ -1,7 +1,7 @@
-import time
 import contextlib
 import logging
 
+from .triggers import TimeTrigger
 from .exceptions import ConfigurationError, DuplicateFieldError
 
 logger = logging.getLogger(__name__)
@@ -94,38 +94,3 @@ class Controller(object):
                 # Log exception as error, rather than exception for simpler
                 # log message. Continue afterwards.
                 logger.error('Failed to update %s because of %r', t, e)
-
-
-class TimeTrigger(object):
-    def __init__(self, timefcn, delayfcn):
-        self._timefcn = timefcn if timefcn else time.time
-        self._delayfcn = delayfcn if delayfcn else time.sleep
-        self._num_steps = 1
-        self._time_step = 0
-
-    def set_steps(self, time_step, num_steps):
-        self._time_step = time_step
-        self._num_steps = num_steps
-
-    def iter(self):
-        logger.info('Starting iterator')
-
-        time_finish_now = self._timefcn()
-        time_start_next = time_finish_now
-
-        for count_steps in range(self._num_steps):
-            if time_start_next > time_finish_now:
-                self._delayfcn(time_start_next - time_finish_now)
-
-            logger.info('Start iteration #%d', count_steps)
-            yield time_start_next
-
-            time_finish_now = self._timefcn()
-            if self._time_step > 0:
-                time_start_next += self._time_step
-                while time_finish_now > time_start_next:
-                    time_start_next += self._time_step
-            else:
-                time_start_next = time_finish_now
-
-        logger.info('Finished iterator')
